Contratos realizados, excepto update

+++ src/cu/edu/cujae/dto/ContractDTO.java
@@ -104,7 +104,7 @@ public class ContractDTO {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(float value) {
         this.value = value;
     }
